+++ changelog.md
@@ -1,8 +1,10 @@
 # Changelog
 
-## v.2.1.2.5
+## v.2.1.2.6
 
-* [#157](https://github.com/PSKeePass/PoShKeePass/issues/157) - Set New-KPConnection function back to internal function and no longer exports.
+* [#158](https://github.com/PSKeePass/PoShKeePass/issues/158) - Added Update-KeePassDatabaseConfiguration function and tests.
+
+* v.2.1.2.5 - Fix - [#157](https://github.com/PSKeePass/PoShKeePass/issues/157) - Set New-KPConnection function back to internal function and no longer exports.
 
 ## v.2.1.2.4